% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_arcpy_env.R
\name{is_arcpy_config}
\alias{is_arcpy_config}
\title{Verifica si arcpy está disponible}
\usage{
is_arcpy_config()
}
\value{
Un valor lógico: \code{TRUE} si \code{arcpy} está disponible, \code{FALSE} en caso contrario.
}
\description{
Comprueba si el módulo \code{arcpy} de ArcGIS está configurado y accesible
en el entorno de Python que \code{reticulate} está utilizando actualmente.
}
