% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_trans_sup_ha.R
\name{st_trans_sup_ha}
\alias{st_trans_sup_ha}
\title{Procesar y calcular superficie en hectáreas para una capa espacial}
\usage{
st_trans_sup_ha(
  shapefile_path,
  campo_superficie = "Sup_ha",
  precision = 8L,
  escala = 2L
)
}
\description{
Esta función es un envoltorio (wrapper) para una herramienta de geoprocesamiento en Python
para preservar el orden de los campos y aplica el formato correcto. Finalmente,
calcula o recalcula el área.
}
\section{Configuración del Entorno}{

Antes de usar esta función, es \strong{imprescindible} configurar el entorno de Python
para que apunte a la instalación de ArcGIS Pro. Utilice la función de ayuda \code{setup_arcpy_env()}:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Ejecutar una vez por sesión de R
setup_arcpy_env()
}\if{html}{\out{</div>}}

Se recomienda ejecutar este comando al inicio de tu script o sesión de R.
}

\examples{
\dontrun{
  # Paso 1: Configurar el entorno de Python con arcpy
  setup_arcpy_env()
  
  # Paso 2: Ejecutar la función en una capa
  ruta_capa <- "C:/ruta/a/tus/datos/mi_capa.shp"
  st_trans_sup_ha(shapefile_path = ruta_capa)

  # También se puede usar con lapply para procesar una lista de capas
  lista_capas <- list.files("C:/ruta/a/tus/datos/", pattern = ".shp$", full.names = TRUE)
  lapply(lista_capas, st_trans_sup_ha)
}
}
